requirements.txt
torch
torchaudio
librosa
numpy
dtw-python
datasets
transformers
pandas
jiwer
elevenlabs